Fix Preview and Simulator weirdness on Xcode 12
The VALID_ARCHS flag was formerly a part of the regular build settings. Xcode now discourages their use (https://stackoverflow.com/a/63259096/7138792) and automatically converted it to a user-defined flag in Xcode 12. Removing it fixes the issues. (I think we had it in there to fix an issue with a prior Xcode beta.)